testing...
In playing around with React and Spring, I realized that it would be unwise
to force the Spring application to serve the React files. I have not played with
React before, and I always thought that it required another backend system to serve the
files...

The TicTacToeController was going to serve the react files. I was testing how Spring sends files.
Static files do not need to be handled through the controllers. The implementation I was playing with
is missing Thyme. I will remove this controller and move back to the React app next.

I may use the Spring app to keep a record of the games in some way, or maybe use it to authinticate
users. Even better would be to use it set up games with multiple users.